+++ exoPoo/src/main/java/org/example/tpBibliotheque/NumericBook.java
@@ -2,9 +2,9 @@ package org.example.tpBibliotheque;
 
 public class NumericBook extends Book{
     private String format;
-    private long size;
+    private double size;
 
-    public NumericBook(String title, String author, String format, long size) {
+    public NumericBook(String title, String author, String format, double size) {
         super(title, author);
         this.format = format;
         this.size = size;
